agent/pyproject.toml
[tool.poetry]
name = "github_inquirer"
version = "0.1.0"
license = "MIT"
description = "agent"
authors = ["ArnavK-09 <arnavkaushal09@gmail.com"]

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[project]
name = "github_inquirer"
version = "0.0.1"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = "^3.12"
langchain-anthropic = "^0.2.1"
langchain = "^0.3.1"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "0.1.30"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
langchain-core = "^0.3.25"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}
langchain-groq = "^0.2.2"
langchain-google-genai = "^2.0.7"

[tool.poetry.scripts]
agent = "github_inquirer.api:main"
